+++ src/command.ts
@@ -1,4 +1,4 @@
-import { ExtensionContext, commands, window, env } from "vscode";
+import { ExtensionContext, commands, window, env, ConfigurationTarget } from "vscode";
 import { Commands, ConfigurationKeys } from "./types";
 import { setting } from "./setting";
 import { getIP } from "./sysinfo";
@@ -13,7 +13,7 @@ class Command {
           if (ip) {
             env.clipboard.writeText(ip);
             window.showInformationMessage(`The IP ${ip} was copied successfully`);
-          }else{
+          } else {
             window.showInformationMessage(`Not found IP`);
           }
         } catch (err) {
@@ -26,7 +26,7 @@ class Command {
     commands.registerCommand(
       Commands.EnableAll,
       () => {
-        setting.cfg?.update(ConfigurationKeys.AllEnabled, true, true);
+        setting.cfg?.update(ConfigurationKeys.AllEnabled, true, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -34,7 +34,7 @@ class Command {
     commands.registerCommand(
       Commands.DisableAll,
       () => {
-        setting.cfg?.update(ConfigurationKeys.AllEnabled, false, true);
+        setting.cfg?.update(ConfigurationKeys.AllEnabled, false, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -42,7 +42,7 @@ class Command {
     commands.registerCommand(
       Commands.EnableCpuLoad,
       () => {
-        setting.cfg?.update(ConfigurationKeys.CpuLoadEnabled, true, true);
+        setting.cfg?.update(ConfigurationKeys.CpuLoadEnabled, true, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -50,7 +50,7 @@ class Command {
     commands.registerCommand(
       Commands.DisableCpuLoad,
       () => {
-        setting.cfg?.update(ConfigurationKeys.CpuLoadEnabled, false, true);
+        setting.cfg?.update(ConfigurationKeys.CpuLoadEnabled, false, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -58,7 +58,7 @@ class Command {
     commands.registerCommand(
       Commands.EnableLoadavg,
       () => {
-        setting.cfg?.update(ConfigurationKeys.LoadavgEnabled, true, true);
+        setting.cfg?.update(ConfigurationKeys.LoadavgEnabled, true, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -66,7 +66,7 @@ class Command {
     commands.registerCommand(
       Commands.DisableLoadavg,
       () => {
-        setting.cfg?.update(ConfigurationKeys.LoadavgEnabled, false, true);
+        setting.cfg?.update(ConfigurationKeys.LoadavgEnabled, false, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -74,7 +74,7 @@ class Command {
     commands.registerCommand(
       Commands.EnableNetworkSpeed,
       () => {
-        setting.cfg?.update(ConfigurationKeys.NetworkSpeedEnabled, true, true);
+        setting.cfg?.update(ConfigurationKeys.NetworkSpeedEnabled, true, ConfigurationTarget.Workspace);
       },
       this
     );
@@ -82,7 +82,39 @@ class Command {
     commands.registerCommand(
       Commands.DisableNetworkSpeed,
       () => {
-        setting.cfg?.update(ConfigurationKeys.NetworkSpeedEnabled, false, true);
+        setting.cfg?.update(ConfigurationKeys.NetworkSpeedEnabled, false, ConfigurationTarget.Workspace);
+      },
+      this
+    );
+
+    commands.registerCommand(
+      Commands.EnableMemoUsage,
+      () => {
+        setting.cfg?.update(ConfigurationKeys.MemoUsageEnabled, true, ConfigurationTarget.Workspace);
+      },
+      this
+    );
+
+    commands.registerCommand(
+      Commands.DisableMemoUsage,
+      () => {
+        setting.cfg?.update(ConfigurationKeys.MemoUsageEnabled, false, ConfigurationTarget.Workspace);
+      },
+      this
+    );
+
+    commands.registerCommand(
+      Commands.EnableUptime,
+      () => {
+        setting.cfg?.update(ConfigurationKeys.UptimeEnabled, true, ConfigurationTarget.Workspace);
+      },
+      this
+    );
+
+    commands.registerCommand(
+      Commands.DisableUptime,
+      () => {
+        setting.cfg?.update(ConfigurationKeys.UptimeEnabled, false, ConfigurationTarget.Workspace);
       },
       this
     );