+++ src/sysinfo.ts
@@ -1,25 +1,24 @@
 import * as si from "systeminformation";
 import * as os from "os";
-import { StatsModules } from "./types";
 
 export async function getCpuSpeed() {
   try {
     const res = await si.cpuCurrentSpeed();
-    return `${res.avg}`;
+    return res.avg;
   } catch (err) {}
 }
 
 export async function getCpuLoad() {
   try {
     const res = await si.currentLoad();
-    return `${res.currentLoad.toFixed(0)}%`;
+    return res.currentLoad;
   } catch (err) {}
 }
 
-export async function getLoad() {
+export async function getLoadavg() {
   try {
     const res = os.loadavg();
-    return res.map((item) => item.toFixed(2)).join(", ");
+    return res;
   } catch (err) {}
 }
 
@@ -35,29 +34,42 @@ export async function getNetworkSpeed() {
     const defaultInterface = await si.networkInterfaceDefault();
     const res = await si.networkStats(defaultInterface);
     const cur = res[0];
-    return `U: ${formatNetowk(cur.tx_sec)} - D: ${formatNetowk(cur.rx_sec)}`;
+    return {
+      up: cur.tx_sec,
+      down: cur.rx_sec,
+    };
   } catch (err) {}
 }
 
-function formatNetowk(data: number) {
-  const KB = 1024;
-  const MB = 1024 * 1024;
-  const GB = 1024 * 1024 * 1024;
-  let res;
-  if (data < KB) {
-    res = 0 + "KB/s";
-  } else if (data < MB) {
-    res = (data / KB).toFixed(0) + "KB/s";
-  } else if (data < GB) {
-    res = (data / MB).toFixed(0) + "MB/s";
-  } else {
-    res = (data / GB).toFixed(0) + "GB/s";
-  }
-  return res;
+export async function getUpTime() {
+  try {
+    return os.uptime();
+  } catch (err) {}
 }
 
-export const sysinfoData: { [key in StatsModules]: () => Promise<string | undefined> } = {
+export async function getMemoryUsage() {
+  try {
+    const res = await si.mem();
+    return {
+      total: res.total,
+      free: res.free,
+      used: res.used,
+      active: res.active,
+      swapTotal: res.swaptotal,
+      swapUsed: res.swapused,
+      swapFree: res.swapfree,
+    };
+  } catch (err) {}
+}
+
+export const sysinfoData = {
   cpuLoad: getCpuLoad,
-  loadavg: getLoad,
+  loadavg: getLoadavg,
   networkSpeed: getNetworkSpeed,
+  memoUsage: getMemoryUsage,
+  uptime: getUpTime,
 };
+
+export type SysinfoData = typeof sysinfoData;
+
+export type StatsModules = keyof typeof sysinfoData;