+++ src/types.ts
@@ -1,16 +1,18 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-export type StatsModules = "cpuLoad" | "loadavg" | "networkSpeed";
-
 export enum Commands {
   "CopyIp" = "StatsBar.CopyIp",
   "EnableAll" = "StatsBar.EnableAll",
   "DisableAll" = "StatsBar.DisableAll",
   "EnableCpuLoad" = "StatsBar.EnableCpuLoad",
-  "EnableNetworkSpeed" = "StatsBar.EnableNetworkSpeed",
-  "EnableLoadavg" = "StatsBar.EnableLoadavg",
   "DisableCpuLoad" = "StatsBar.DisableCpuLoad",
+  "EnableNetworkSpeed" = "StatsBar.EnableNetworkSpeed",
   "DisableNetworkSpeed" = "StatsBar.DisableNetworkSpeed",
+  "EnableLoadavg" = "StatsBar.EnableLoadavg",
   "DisableLoadavg" = "StatsBar.DisableLoadavg",
+  "EnableMemoUsage" = "StatsBar.EnableMemoUsage",
+  "DisableMemoUsage" = "StatsBar.DisableMemoUsage",
+  "EnableUptime" = "StatsBar.EnableUptime",
+  "DisableUptime" = "StatsBar.DisableUptime",
 }
 
 export enum ConfigurationKeys {
@@ -18,6 +20,8 @@ export enum ConfigurationKeys {
   CpuLoadEnabled = "cpuLoad.enabled",
   LoadavgEnabled = "loadavg.enabled",
   NetworkSpeedEnabled = "networkSpeed.enabled",
+  MemoUsageEnabled = "memoUsage.enabled",
+  UptimeEnabled = "uptime.enabled",
   RefreshInterval = "refreshInterval",
   Location = "location",
 }