+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the stats bar extension will be documented in this file.
 
+## [0.2.0] - 2021-08-28
+
+### Added
+
+- Support show memory usage and uptime
+
+### Fixed
+
+- Separate configurations for different workspaces
+
 ## [0.1.0] - 2021-08-28
 
 ### Added